pyproject.toml
[project]
name = "dnd-dice-chat"
version = "0.1.0"
description = "D&D Dice Chat backend service"
authors = ["Vladimit Vzyatchinkov <main@nardonist.ru>"]
requires-python = ">=3.11"
dependencies = [
    "alembic==1.16.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==21.2.0",
    "asyncpg==0.30.0",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "click==8.2.1",
    "dnspython==2.7.0",
    "email_validator==2.2.0",
    "fastapi==0.115.13",
    "fastapi-cli==0.0.7",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "Jinja2==3.1.6",
    "Mako==1.3.10",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "mdurl==0.1.2",
    "psycopg2==2.9.10",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic_core==2.33.2",
    "Pygments==2.19.2",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "PyYAML==6.0.2",
    "rich==14.0.0",
    "rich-toolkit==0.14.7",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.41",
    "sqlmodel==0.0.24",
    "starlette==0.46.2",
    "typer==0.16.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.0",
    "uvicorn==0.34.3",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1"
]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
disallow_untyped_defs = true
check_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
show_error_codes = true
pretty = true

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 120

[tool.black]
line-length = 120
